I have made just a few changes to the filepath, to make it easier for the grader to understand where everything is. I assure you no real changes to the code itself was made in this git push.
For the grader, please look at the 'updated_Dir' directory. In it, we have 3 julia scripts-one main, one testing, and one benchmarking. All of these scripts are quite long and produce plots, which can be looked at, but the main requirements are simply outputted in terminal when these scripts are run. The main work of fitting in parallel is done in the main script, which produces all the outputs that we proposed to produce. The testing script compares parallel and serial outputs, confirming that they are the same, and the benchmarking ('timing') script times the two versions of the code.
Again, the purpose of this commit is simply to clarify what exactly is happening-no real changes to the code has been made here. You may check the specifics of the Git commit to confirm this.